products-details

+++ delete_product.php
@@ -7,9 +7,15 @@ if (!isset($_SESSION['user_id'])) {
     exit();
 }
 
+$user_id = $_SESSION['user_id'];
+
+if (!isset($_POST['id'])) {
+    die("Invalid Request");
+}
+
 $id = $_POST['id'];
 
-mysqli_query($conn, "delete from products where product_id='$id'");
+mysqli_query($conn, "delete from products where product_id='$id' and user_id='$user_id'");
 
 header("Location: products.php");
 exit();
